kafka_prod/python-publisher/requirements.txt
essential_generators
kafka-python